+++ packages/picker-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 8.2.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@12.4.7
+
 ## 8.2.10
 
 ### Patch Changes